Less code duplications
