+++ miniapps/bicgstab.hh
@@ -63,14 +63,14 @@ std::tuple<double, int> BiCGSTAB<T>::apply(const spblas::csr_view<T> a,
   // prev_rho = rho = omega = alpha = beta = gamma = 1.0
   // rr = v = s = t = z = y = p = 0
   std::vector<T> r(b);
-  std::vector<T> rr(b.size(), 0);
-  std::vector<T> v(b.size(), 0);
-  std::vector<T> s(b.size(), 0);
-  std::vector<T> t(b.size(), 0);
-  std::vector<T> z(b.size(), 0);
-  std::vector<T> y(b.size(), 0);
-  std::vector<T> p(b.size(), 0);
-  std::vector<T> tmp(b.size(), 0);
+  std::vector<T> rr(b.size(), 0.0);
+  std::vector<T> v(b.size(), 0.0);
+  std::vector<T> s(b.size(), 0.0);
+  std::vector<T> t(b.size(), 0.0);
+  std::vector<T> z(b.size(), 0.0);
+  std::vector<T> y(b.size(), 0.0);
+  std::vector<T> p(b.size(), 0.0);
+  std::vector<T> tmp(b.size(), 0.0);
 
   double prev_rho, rho, omega, alpha, beta, gamma;
   prev_rho = rho = omega = alpha = beta = gamma = 1.0;
@@ -83,7 +83,7 @@ std::tuple<double, int> BiCGSTAB<T>::apply(const spblas::csr_view<T> a,
   while (true) {
     iters++;
 
-    // rho = dot(r, z)
+    // rho = dot(rr, r)
     rho = blas::dot(r.size(), rr.data(), 1, r.data(), 1);
 
     error = blas::nrm2(r.size(), r.data(), 1);
@@ -95,11 +95,11 @@ std::tuple<double, int> BiCGSTAB<T>::apply(const spblas::csr_view<T> a,
       break;
     }
 
-    if (iters == 1) {
+    if (prev_rho * omega == 0.0) {
       p = r;
     } else {
       // beta = (rho / prev_rho) * (alpha / omega)
-      beta = (rho / prev_rho) * (alpha / omega)
+      beta = (rho / prev_rho) * (alpha / omega);
       // p = r + beta * (p - omega * v)
       tmp = p;
       blas::axpy(tmp.size(), -omega, v.data(), 1, tmp.data(), 1);
@@ -112,13 +112,19 @@ std::tuple<double, int> BiCGSTAB<T>::apply(const spblas::csr_view<T> a,
     y = p;
     // v = A * y
     spblas::multiply(a, y, v);
-    // alpha = rho / dot(rr, v)
-    alpha = rho / blas::dot(rr.size(), rr.data(), 1, v.data(), 1);
-    // s = r - alpha * v
-    tmp = v;
-    blas::scal(tmp.size(), -alpha, tmp.data(), 1);
-    blas::axpy(tmp.size(), 1.0, r.data(), 1, tmp.data(), 1);
-    s = tmp;
+    // beta = dot(rr, v)
+    beta = blas::dot(v.size(), rr.data(), 1, v.data(), 1);
+    if (beta == 0.0) {
+      s = r;
+    } else {
+      // alpha = rho / beta
+      alpha = rho / beta;
+      // s = r - alpha * v
+      tmp = v;
+      blas::scal(tmp.size(), -alpha, tmp.data(), 1);
+      blas::axpy(tmp.size(), 1.0, r.data(), 1, tmp.data(), 1);
+      s = tmp;
+    }
 
     error = blas::nrm2(s.size(), s.data(), 1);
     if (iters >= max_iters || error < eps) {
@@ -126,6 +132,7 @@ std::tuple<double, int> BiCGSTAB<T>::apply(const spblas::csr_view<T> a,
       std::cout << "iters = " << iters << "\n";
       std::cout << "eps = " << eps << "\n";
       std::cout << "max_iters = " << max_iters << "\n";
+      blas::axpy(x.size(), alpha, y.data(), 1, x.data(), 1);
       break;
     }
 
@@ -133,12 +140,16 @@ std::tuple<double, int> BiCGSTAB<T>::apply(const spblas::csr_view<T> a,
     z = s;
     // t = A * z
     spblas::multiply(a, z, t);
-    // gamma = dot(s, t)
+    // gamma = dot(t, s)
     gamma = blas::dot(s.size(), t.data(), 1, s.data(), 1);
     // beta = dot(t, t)
     beta = blas::dot(t.size(), t.data(), 1, t.data(), 1);
     // omega = gamma / beta
-    omega = gamma / beta;
+    if (beta == 0.0) {
+      omega = 0.0;
+    } else {
+      omega = gamma / beta;
+    }
     // x = x + alpha * y + omega * z
     tmp = z;
     blas::scal(tmp.size(), omega, tmp.data(), 1);
@@ -151,7 +162,7 @@ std::tuple<double, int> BiCGSTAB<T>::apply(const spblas::csr_view<T> a,
     blas::axpy(tmp.size(), 1.0, s.data(), 1, tmp.data(), 1);
     r = tmp;
 
-    std::swap(rho, prev_rho);
+    std::swap(prev_rho, rho);
   }
   return {error, iters};
 }
